fix Bard unit tests
